requirements.txt
flask
flask-socketio
eventlet
diffusers
transformers
torch
torchvision
torchaudio
Pillow
huggingface_hub